+++ draft-js-static-toolbar-plugin/src/components/Separator/index.js
@@ -1,6 +1,11 @@
 import React from 'react';
-import separatorStyles from '../../separatorStyles.css';
+import { css } from 'linaria';
 
-export default ({ className = separatorStyles.separator }) => (
-  <div className={className} />
-);
+const separator = css`
+  display: inline-block;
+  border-right: 1px solid #ddd;
+  height: 24px;
+  margin: 0 0.5em;
+`;
+
+export default ({ className = separator }) => <div className={className} />;