+++ draft-js-static-toolbar-plugin/CHANGELOG.md
@@ -3,15 +3,21 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## To be released
+
+- Migrate styles to linaria
+
 ### 3.0.1
 
 - Allow draft-js v0.11
 - Remove unused dependencies
 
 ### 3.0.0
+
 - Swap structure prop in static toolbar for render prop
 
 ## 2.0.2 - 2.0.3
+
 - bumped find-with-regex
 
 ## To Be Released